support mermaid diagram
