+++ src/vite-env.d.ts
@@ -1,1 +1,8 @@
 /// <reference types="vite/client" />
+
+declare module '@liradb2000/markdown-it-mermaid' {
+  import { PluginSimple } from 'markdown-it';
+
+  const markdownItMermaid: PluginSimple;
+  export default markdownItMermaid;
+}
